+++ tests/psp22_contract/lib.rs
@@ -1,0 +1,151 @@
+#![cfg_attr(not(feature = "std"), no_std, no_main)]
+
+// An example code of a smart contract using the PSP22Data struct to implement
+// the functionality of a PSP22 fungible token.
+//
+// Copied from the PSP22 repository and modified.
+#[ink::contract]
+mod token {
+    use ink::prelude::vec::Vec;
+    use psp22::{PSP22Data, PSP22Error, PSP22Event, PSP22};
+
+    #[ink(storage)]
+    pub struct Token {
+        data: PSP22Data,
+    }
+
+    impl Token {
+        #[ink(constructor)]
+        pub fn new(supply: u128) -> Self {
+            Self {
+                data: PSP22Data::new(supply, Self::env().caller()),
+            }
+        }
+
+        #[ink(message)]
+        pub fn burn(&mut self, amount: u128) -> Result<(), PSP22Error> {
+            let events = self.data.burn(self.env().caller(), amount)?;
+            self.emit_events(events);
+            Ok(())
+        }
+
+        fn emit_events(&self, events: Vec<PSP22Event>) {
+            for event in events {
+                match event {
+                    PSP22Event::Transfer { from, to, value } => {
+                        self.env().emit_event(Transfer { from, to, value })
+                    }
+                    PSP22Event::Approval {
+                        owner,
+                        spender,
+                        amount,
+                    } => self.env().emit_event(Approval {
+                        owner,
+                        spender,
+                        amount,
+                    }),
+                }
+            }
+        }
+    }
+
+    #[ink(event)]
+    pub struct Approval {
+        #[ink(topic)]
+        owner: AccountId,
+        #[ink(topic)]
+        spender: AccountId,
+        amount: u128,
+    }
+
+    #[ink(event)]
+    pub struct Transfer {
+        #[ink(topic)]
+        from: Option<AccountId>,
+        #[ink(topic)]
+        to: Option<AccountId>,
+        value: u128,
+    }
+
+    impl PSP22 for Token {
+        #[ink(message)]
+        fn total_supply(&self) -> u128 {
+            self.data.total_supply()
+        }
+
+        #[ink(message)]
+        fn balance_of(&self, owner: AccountId) -> u128 {
+            self.data.balance_of(owner)
+        }
+
+        #[ink(message)]
+        fn allowance(&self, owner: AccountId, spender: AccountId) -> u128 {
+            self.data.allowance(owner, spender)
+        }
+
+        #[ink(message)]
+        fn transfer(
+            &mut self,
+            to: AccountId,
+            value: u128,
+            _data: Vec<u8>,
+        ) -> Result<(), PSP22Error> {
+            let events = self.data.transfer(self.env().caller(), to, value)?;
+            self.emit_events(events);
+            Ok(())
+        }
+
+        #[ink(message)]
+        fn transfer_from(
+            &mut self,
+            from: AccountId,
+            to: AccountId,
+            value: u128,
+            _data: Vec<u8>,
+        ) -> Result<(), PSP22Error> {
+            let events = self
+                .data
+                .transfer_from(self.env().caller(), from, to, value)?;
+            self.emit_events(events);
+            Ok(())
+        }
+
+        #[ink(message)]
+        fn approve(&mut self, spender: AccountId, value: u128) -> Result<(), PSP22Error> {
+            let events = self.data.approve(self.env().caller(), spender, value)?;
+            self.emit_events(events);
+            Ok(())
+        }
+
+        #[ink(message)]
+        fn increase_allowance(
+            &mut self,
+            spender: AccountId,
+            delta_value: u128,
+        ) -> Result<(), PSP22Error> {
+            let events = self
+                .data
+                .increase_allowance(self.env().caller(), spender, delta_value)?;
+            self.emit_events(events);
+            Ok(())
+        }
+
+        #[ink(message)]
+        fn decrease_allowance(
+            &mut self,
+            spender: AccountId,
+            delta_value: u128,
+        ) -> Result<(), PSP22Error> {
+            let events = self
+                .data
+                .decrease_allowance(self.env().caller(), spender, delta_value)?;
+            self.emit_events(events);
+            Ok(())
+        }
+    }
+
+    #[cfg(test)]
+    mod tests {
+        crate::tests!(Token, (|supply| Token::new(supply, None, None, 0)));
+    }
+}