+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `ExecCall` gets new trait bound `T` to reperesent type of the returned value.
+
+### Added
+
+- Support for `DRink!` added to `ink-wrapper-types`. Currently fixed to work only with `MinimalRuntime`.
+
+### Removed
+
+- Support for `aleph-client` is removed from `ink-wrapper-types`. This is due to the fact that `aleph-client` is not published to crates.io anymore, which would make it impossible to publish `ink-wrapper-types` as well. Support for live chains will come back in the future, by the use of `subxt` as the backend.
+
 ## 0.6.1
 
 ### Changed